모델 코드 여러 파일로 분리
