--- BoxOffice/Model/Model.swift
@@ -1,97 +1,0 @@
-//
-//  Model.swift
-//  BoxOffice
-//
-//  Created by LEE JUNSANG on 2018. 12. 16..
-//  Copyright © 2018년 zunzun. All rights reserved.
-//
-
-import Foundation
-
-struct APIResponse: Codable {
-    let orderType: Int
-    let movies: [Movie]
-    
-    enum CodingKeys: String, CodingKey {
-        case movies
-        case orderType = "order_type"
-    }
-}
-
-struct Movie: Codable {
-    let title: String
-    let date: String
-    let thumb: String
-    let reservationGrade: Int
-    let userRating: Double
-    let grade: Int
-    let reservationRate: Double
-    let id: String
-    
-    enum CodingKeys: String, CodingKey {
-        case title, date, thumb, grade, id
-        case reservationGrade = "reservation_grade"
-        case userRating = "user_rating"
-        case reservationRate = "reservation_rate"
-    }
-}
-
-struct MovieDetailInfo: Codable {
-    let audience: Int
-    let actor: String
-    let duration: Int
-    let director: String
-    let synopsis: String
-    let genre: String
-    let grade: Int
-    let image: String
-    let reservationGrade: Int
-    let title: String
-    let reservationRate: Double
-    let userRating: Double
-    let date: String
-    let id: String
-    
-    var audienceString: String {
-        var value = audience
-        var temp: String = ""
-        while true {
-            if value / 1000 == 0 {
-                return "\(value % 1000)\(temp)"
-            } else {
-                temp = ",\(value % 1000)\(temp)"
-                value = value / 1000
-            }
-        }
-    }
-    
-    enum CodingKeys: String, CodingKey {
-        case audience, actor, duration, director, synopsis, genre, grade, image, title, date, id
-        case reservationGrade = "reservation_grade"
-        case reservationRate = "reservation_rate"
-        case userRating = "user_rating"
-    }
-}
-
-struct CommentsList: Codable {
-    let movieId: String
-    let comments: [Comment]
-    
-    enum CodingKeys: String, CodingKey {
-        case comments
-        case movieId = "movie_id"
-    }
-}
-
-struct Comment: Codable {
-    let rating: Double
-    let timestamp: Double
-    let writer: String
-    let movieId: String
-    let contents: String
-    
-    enum CodingKeys: String, CodingKey {
-        case rating, timestamp, writer, contents
-        case movieId = "movie_id"
-    }
-}
